+++ src/FaluSdk/Files/FilesListOptions.cs
@@ -1,13 +1,13 @@
 ﻿using Falu.Core;
 using System.Collections.Generic;
 
-namespace Falu.FileUploads
+namespace Falu.Files
 {
-    /// <summary>Options for filtering and pagination of list file uploads.</summary>
-    public record FileUploadsListOptions : BasicListOptions
+    /// <summary>Options for filtering and pagination of list files.</summary>
+    public record FilesListOptions : BasicListOptions
     {
         /// <summary>Filter options for <code>purpose</code> property.</summary>
-        public List<FileUploadPurpose>? Purpose { get; set; }
+        public List<FilePurpose>? Purpose { get; set; }
 
         /// <inheritdoc/>
         internal override void Populate(QueryValues values)
